fix: ASCII table CSS styling affecting other icons on page
The styling from was indavertantly colouring other icons on the page, when it should have only been modifying the colours of the symbols in the SVGs.

This has been fixed by grouping the individual paths that make up the symbols into a `symbols` group, ensuring that its ID is set to `symbols` in the object properties in Inkscape, clearing existing CSS styling from the SVG and updating the CSS selector in `src/css/custom.css` to only modify the `symbols` group.