feat: enable themed favicon
The colour of the logo in the favicon is determined by a stylesheet in the SVG itself - it uses `prefers-color-scheme` in order to colour the logo.

It does not follow the theme of the website but follows the browser/OS choice.

The SVG in this commit had to be manually edited, since Inkscape didn't want to play nice with the CSS. If this file gets edited in the future, be careful to not accidentally break this functionality.

Using an SVG here was preferred over using the existing `favicon-white.png` and `favicon.ico` because Docusaurus did not provide a handy feature to toggle between the light/dark mode favicons, in the same way that it did for the logo in the navbar (with `navbar.logo.src` and `navbar.logo.srcDark`).